move schema generation to FuzzDataGenerator
